+++ react-ui/src/components/PracticeResult/PracticeResult.js
@@ -1,0 +1,59 @@
+import React, { Component } from 'react';
+import styled from 'styled-components';
+
+const Wrapper = styled.div`
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+`;
+
+const Header = styled.div`
+  margin-top: 20px;
+  font-family: 'Open Sans';
+  font-size: 35px;
+  font-weight: 600;
+  color: #454545;
+`;
+
+class PracticeResult extends Component {
+
+  _avatar() {
+    const funnyFaces = [
+      'https://s3.eu-west-2.amazonaws.com/chineseme/personalities/biaoqing/marvin_question.png',
+      'https://s3.eu-west-2.amazonaws.com/chineseme/personalities/biaoqing/yuguo_happy.png',
+      'https://s3.eu-west-2.amazonaws.com/chineseme/personalities/biaoqing/liyu_surprised.png',
+      'https://s3.eu-west-2.amazonaws.com/chineseme/personalities/biaoqing/meixin_surprised.png'
+    ];
+    return funnyFaces[Math.floor(Math.random() * funnyFaces.length)];
+  }
+
+  _message() {
+    const messages = [
+      'Good job!',
+      'Well done!',
+      'Great. Never stop learning!',
+      'You\'re doing well!'
+    ];
+    return messages[Math.floor(Math.random() * messages.length)];
+  }
+
+  render() {
+    return (
+      <Wrapper>
+        <img
+          src={this._avatar()}
+          alt=""
+          width="175"
+          height="175"
+        />
+        <Header><i>{ this._message() }</i></Header>
+      </Wrapper>
+    );
+  }
+}
+
+PracticeResult.propTypes = {};
+
+export default PracticeResult;