+++ react-ui/src/sagas/study/practice.js
@@ -1,9 +1,10 @@
 import { put, select, call, take, race } from 'redux-saga/effects';
+import { delay } from 'redux-saga';
 import { actions as uiActions } from '../../redux/ui';
 import selectors from '../../rootSelectors';
 import { fetchEntities } from '../entities';
 import { actions as practiceActions } from '../../redux/practice';
-import { types as sagaTypes } from '../actions';
+import { types as sagaTypes, actions as sagaActions } from '../actions';
 import getStudyFunctions from '../../helpers/getStudyFunctions';
 import mapExerciseTypeToSetCurrentAction
   from '../../helpers/mapExerciseTypeToSetCurrentAction';
@@ -68,6 +69,10 @@ export function* run() {
       completed = true;
     }
   }
+  // yield delay(1000);
+  yield put(sagaActions.playLevelWinSound());
+  yield put(uiActions.set('nextButton', true));
+  yield take(sagaTypes.NEXT);
 }
 
 export function* nextScreen() {