+++ CHANGELOG.md
@@ -4,15 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1] 2017-12-XX
+## [0.1] 2017-12-05
 
 ### Added
 - Auto-capitalize user's family and given names before saving them in our DB.
+- Links in TOC (using react-sroll)
+- Kind success messages at the end of exercises
 
 ### Content
 - S0E3: Separate sound buttons for 姓 and 叫
 - S0E3: "I am Chinese" should be in italic
 - S0E2: Dialogs should mention the user's full name instead of given name only.
+- S0E2: "Example dialogs" now only have an audio button.
 
 ### Changed
 - Dialogs start with "Explore" when opened (previously "Listen").
@@ -24,3 +27,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Stroke order exercise in Review say "Write the character" instead of "Your turn!".
 
 ### Fixed
+- Robotic voice no longer pronounce the underscore symbols.